requirements.txt
streamlit
chromadb
sentence-transformers
anthropic
python-dotenv
beautifulsoup4
requests
pdfplumber
youtube-transcript-api
pyyaml
